move field errors into a separate module
