+++ landlab/__init__.py
@@ -17,7 +17,7 @@ from ._version import get_versions
 from .core.errors import MissingKeyError, ParameterValueError
 from .core.model_component import Component
 from .core.model_parameter_loader import load_params
-from .field.scalar_data_fields import FieldError
+from .field import FieldError
 from .grid import (
     HexModelGrid,
     ModelGrid,
